Remove: unused imports and commented out code.Rename: Form and Signin components.
