--- data/tempFacts.js
@@ -1,24 +1,0 @@
-export default {
-  data: [
-    {
-      fact: 'Mountain lions are strong jumpers, thanks to muscular hind legs that are longer than their front legs.',
-      length: 102,
-    },
-    {
-      fact: 'The smallest pedigreed cat is a Singapura, which can weigh just 4 lbs (1.8 kg), or about five large cans of cat food. The largest pedigreed cats are Maine Coon cats, which can weigh 25 lbs (11.3 kg), or nearly twice as much as an average cat weighs.',
-      length: 249,
-    },
-    {
-      fact: 'In 1987 cats overtook dogs as the number one pet in America.',
-      length: 60,
-    },
-    // {
-    //   fact: "Like birds, cats have a homing ability that uses its biological clock, the angle of the sun, and the Earth's magnetic field. A cat taken far from its home can return to it. But if a cat's owners move far from its home, the cat can't find them.",
-    //   length: 243,
-    // },
-    // {
-    //   fact: 'The group of words associated with cat (catt, cath, chat, katze) stem from the Latin catus, meaning domestic cat, as opposed to feles, or wild cat.',
-    //   length: 147,
-    // },
-  ],
-};
